5/10 YOon

+++ SpringBoot_caRing/src/main/java/com/example/caRing/model/board/Board.java
@@ -13,8 +13,8 @@ public class Board {
    private String board_contents;
    private LocalDateTime created_time;
    private Long price;
-   private Long lat;
-   private Long lng;
+   private Double lat;
+   private Double lng;
    private Long carInfo_id;
    private String rent_start;
    private String rent_end;
